backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
tensorflow==2.16.1
numpy>=1.24.0
Pillow>=10.0.0
gunicorn==22.0.0
